#0: address more comments
