+++ tt_metal/distributed/config/TG.json
