+++ tt_metal/distributed/system_mesh.cpp
@@ -5,56 +5,12 @@
 #include "tt_metal/distributed/system_mesh.hpp"
 
 #include "umd/device/types/cluster_descriptor_types.h"
+#include "tt_metal/distributed/coordinate_translation.hpp"
 
 namespace tt::tt_metal::distributed {
 
-// TODO: Consider conversion to StrongType instead of alias
-using LogicalCoordinate = Coordinate;
-using PhysicalCoordinate = eth_coord_t;
-using CoordinateTranslationMap = std::unordered_map<LogicalCoordinate, PhysicalCoordinate>;
-
-std::string get_config_path(const std::string& filename) {
-    std::string root_path = getenv("TT_METAL_HOME") ? getenv("TT_METAL_HOME") : "./";
-    return root_path + "/tt_metal/distributed/mesh_configurations/" + filename;
-}
-
-CoordinateTranslationMap load_translation_map(const std::string& filename, const std::string& key) {
-    std::ifstream file(filename);
-    TT_FATAL(!file.is_open(), "Unable to open file: {}", filename);
-
-    nlohmann::json j;
-    try {
-        file >> j;
-    } catch (const nlohmann::json::parse_error& e) {
-        TT_THROW("JSON parsing error in file {}: {}", filename, e.what());
-    }
-
-    TT_FATAL(!j.contains(key), "Key '{}' not found in JSON file: {}", key, filename);
-
-    CoordinateTranslationMap result;
-    for (const auto& mapping : j[key]) {
-        if (mapping.size() != 2 || mapping[0].size() != 2 || mapping[1].size() != 5) {
-            TT_THROW("Invalid coordinate format in JSON file: {}", filename);
-        }
-        result.emplace(
-            LogicalCoordinate{mapping[0][0], mapping[0][1]},
-            PhysicalCoordinate{
-                mapping[1][0],  // cluster_id
-                mapping[1][2],  // x
-                mapping[1][1],  // y
-                mapping[1][3],  // rack
-                mapping[1][4]   // shelf
-            });
-    }
-
-    return result;
-}
-
 class SystemMesh::Impl {
 private:
-    using LogicalCoordinate = Coordinate;
-    using PhysicalCoordinate = eth_coord_t;
-
     std::unordered_map<MeshDeviceID, std::vector<chip_id_t>> assigned_devices_;
     std::unordered_map<MeshDeviceID, std::weak_ptr<MeshDevice>> assigned_mesh_device_devices_;
 
@@ -77,50 +33,8 @@ public:
     IDevice* get_device(const chip_id_t physical_device_id) const;
     void register_mesh_device(const std::shared_ptr<MeshDevice>& mesh_device, const std::vector<IDevice*>& devices);
 
-    static MeshShape get_system_mesh_shape(size_t system_num_devices);
-    static CoordinateTranslationMap get_system_mesh_translation_map(size_t system_num_devices);
-
     chip_id_t get_physical_device_id(size_t logical_row_idx, size_t logical_col_idx) const;
 };
-
-// Implementation of private static methods
-MeshShape SystemMesh::Impl::get_system_mesh_shape(size_t system_num_devices) {
-    static const std::unordered_map<size_t, MeshShape> system_mesh_to_shape = {
-        {1, MeshShape{1, 1}},   // single-device
-        {2, MeshShape{1, 2}},   // N300
-        {8, MeshShape{2, 4}},   // T3000; as ring to match existing tests
-        {32, MeshShape{8, 4}},  // TG, QG
-        {64, MeshShape{8, 8}},  // TGG
-    };
-    TT_FATAL(
-        system_mesh_to_shape.contains(system_num_devices), "Unsupported number of devices: {}", system_num_devices);
-    auto shape = system_mesh_to_shape.at(system_num_devices);
-    log_debug(LogMetal, "Logical SystemMesh Shape: {}x{}", shape.num_rows, shape.num_cols);
-    return shape;
-}
-
-CoordinateTranslationMap SystemMesh::Impl::get_system_mesh_translation_map(size_t system_num_devices) {
-    // TG has 32 non-mmio user devices and 4 mmio devices not exposed to the user
-    // QG has 32 mmio user devices
-    // Once TG is fully deprecated, can remove TG code path
-    std::string galaxy_mesh_descriptor = "TG.json";
-    if (tt::Cluster::instance().number_of_pci_devices() == system_num_devices) {
-        galaxy_mesh_descriptor = "QG.json";
-    }
-    const std::unordered_map<size_t, std::string> system_mesh_translation_map = {
-        {1, "device.json"},
-        {2, "N300.json"},
-        {8, "T3000.json"},
-        {32, galaxy_mesh_descriptor},
-        {64, "TGG.json"},
-    };
-    TT_FATAL(
-        system_mesh_translation_map.contains(system_num_devices),
-        "Unsupported number of devices: {}",
-        system_num_devices);
-    auto translation_config_file = get_config_path(system_mesh_translation_map.at(system_num_devices));
-    return load_translation_map(translation_config_file, "logical_to_physical_coordinates");
-}
 
 // Implementation of public methods
 bool SystemMesh::Impl::is_system_mesh_initialized() const { return !physical_coordinate_to_device_id_.empty(); }
@@ -145,9 +59,7 @@ void SystemMesh::Impl::initialize() {
         }
     }
 
-    auto num_devices = physical_coordinate_to_device_id_.size();
-    logical_mesh_shape_ = get_system_mesh_shape(num_devices);
-    logical_to_physical_coordinates_ = get_system_mesh_translation_map(num_devices);
+    std::tie(logical_to_physical_coordinates_, logical_mesh_shape_) = get_system_mesh_coordinate_translation_map();
     for (const auto& [logical_coordinate, physical_coordinate] : logical_to_physical_coordinates_) {
         logical_to_device_id_.emplace(logical_coordinate, physical_coordinate_to_device_id_.at(physical_coordinate));
     }
